Close path in Man: Beard
This particular path was causing a segfault when building a monochrome font using FontForge. Experimentally, closing the path (which was nearly closed already) seems to prevent this error from occurring.

This change was not included in my main batch of corrections because it does meaningfully (albeit very minorly) alter the content of the src file.