Experiment with Eliminating Transforms
Problem: SVG Cleaner doesn't apply transformations to ellipses or rectangles.
Solution: Convert the objects to paths using Inkscape, and then use the Inkscape Apply-Transforms extension from here:
https://github.com/Klowner/inkscape-applytransforms
Then save as an optimized SVG.

Unfortunately, the resulting SVG needs some adjustment to bring it to the standard OpenMoji SVG format, I ended up pasting the paths into the original files.
And I don't know of a way to automate the process. It might be possible. Inkscape has some sort of command-line interface, but It's unclear to me how to use it.