Update build_black_font_with_fontforge.py
Debugged the skintone variant code. I thought I had eleminated duplicate glyphs, but I had actually just broken things further.
This version has all the skintone variants and still is only 1 MB. Downside: It might not be possible to deal with the duplicate glyphs issue.

This new version of the code takes advantage of the openmoji.csv file.
I was also able to exploit a quirk in that file to find a set of glyphs with official codepoints differing from their filenames.

Also, I tested making an otf instead of a ttf. But the otf is twice the size. I think I do not understand the differences between the two formats' ligature standards.