Remove swimsuit shoulder strap self-intersection
This addresses one of the rendering errors mentioned here:
https://github.com/hfg-gmuend/openmoji/issues/93#issuecomment-1090027056

The cause of this rendering error was that the path which makes up the top of the swimsuit had a self-intersection near the left shoulder-strap. 
The fix was to very slightly adjust one of the coordinates in the curve to remove the self-intersection.
This change will very slightly change the appearance of the rendered image.