Create a Comparison table for the black fonts.
There is a script in the helpers folder which generates the html file in the font folder.

The helpers script reads the openmoji.csv, and creates an html file in the fonts folder.

The html file presents each non-skintone-variant glyph three ways. The first column displays the png from the black/72x72 folder. The second column displays the standard font, built using nanoemoji. The third column displays an alternate font built using fontforge.

The purpose of these comparisons is to easily visually identify glyphs with problematic src files.